refactor(react-components): toggle-button-group のバリアントを index.tsx に統合
MultiToggleButtonGroup / SingleToggleButtonGroup / ToggleButton を
1 ファイル (index.tsx) に併記する形に統合。stories も index.stories.tsx
1 本に集約。barrel index.ts と旧 variant 別ファイルを削除。
