+++ packages/react-components/src/primitives/toggle-button-group/index.stories.tsx
@@ -1,22 +1,48 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { type FC, useId } from "react";
-import { SingleToggleButtonGroup } from "./single-toggle-button-group";
-import { ToggleButton } from "./toggle-button";
+import {
+	MultiToggleButtonGroup,
+	SingleToggleButtonGroup,
+	ToggleButton,
+} from ".";
 
 const meta = {
-	title: "UI/SingleToggleButtonGroup",
-	component: SingleToggleButtonGroup,
+	title: "UI/ToggleButtonGroup",
 	parameters: {
 		layout: "centered",
 	},
 	tags: ["autodocs"],
-} satisfies Meta<typeof SingleToggleButtonGroup>;
+} satisfies Meta;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const NumberOptions: Story = {
-	name: "数値オプション（RPE 5〜10 の 11 段階）",
+export const MultiSelectTextDecoration: Story = {
+	name: "Multi: テキスト装飾（複数選択）",
+	render: () => <MultiTextDecorationDemo />,
+};
+
+const MultiTextDecorationDemo: FC = () => {
+	const baseId = useId();
+	const ids = {
+		bold: `${baseId}-bold`,
+		italic: `${baseId}-italic`,
+		underline: `${baseId}-underline`,
+	};
+	return (
+		<MultiToggleButtonGroup
+			defaultSelectedKeys={[ids.bold, ids.italic]}
+			className="flex gap-1"
+		>
+			<ToggleButton id={ids.bold}>太字</ToggleButton>
+			<ToggleButton id={ids.italic}>斜体</ToggleButton>
+			<ToggleButton id={ids.underline}>下線</ToggleButton>
+		</MultiToggleButtonGroup>
+	);
+};
+
+export const SingleSelectNumberOptions: Story = {
+	name: "Single: 数値オプション（RPE 5〜10 の 11 段階）",
 	render: () => (
 		<SingleToggleButtonGroup
 			defaultSelectedKey="8"
@@ -31,12 +57,12 @@ export const NumberOptions: Story = {
 	),
 };
 
-export const StringOptions: Story = {
-	name: "文字列オプション（テキスト配置）",
-	render: () => <StringOptionsDemo />,
+export const SingleSelectStringOptions: Story = {
+	name: "Single: 文字列オプション（テキスト配置）",
+	render: () => <SingleStringOptionsDemo />,
 };
 
-const StringOptionsDemo: FC = () => {
+const SingleStringOptionsDemo: FC = () => {
 	const baseId = useId();
 	const ids = {
 		left: `${baseId}-left`,
@@ -55,8 +81,8 @@ const StringOptionsDemo: FC = () => {
 	);
 };
 
-export const NoneSelected: Story = {
-	name: "未選択状態",
+export const SingleSelectNoneSelected: Story = {
+	name: "Single: 未選択状態",
 	render: () => (
 		<SingleToggleButtonGroup className="flex flex-wrap gap-1">
 			{[5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 9.5, 10].map((value) => (
