--- packages/react-components/src/primitives/toggle-button-group/index.ts
@@ -1,3 +1,0 @@
-export { MultiToggleButtonGroup } from "./multi-toggle-button-group";
-export { SingleToggleButtonGroup } from "./single-toggle-button-group";
-export { ToggleButton } from "./toggle-button";
